Fix turtle_python exercise
The config.yaml and the config_test.yaml were updated. This update
allows to make use of  the latest version of python-grader-utils.
In addition, the grader_test.py file needed to be updated as well.
The turtles.png.base64 file was added to the turtle_python folder.
This new file avoids the software to crash if an error is found in
the grader_tests validations. The feedback template always tries
to include the turtles.png.base64 template even though the
grader_tests do not generate it if the submission does not pass
the initial validation steps.