Remove the numerical prefix from the file name and Fix links
The numerical prefix was removed from the all the file names. This
change will help to add new chapters by avoiding renaming all the
chapters that would be affected by having to change the numerical
prefix in the file names.

This commit also fixes some links that hard-coded the domain to
`localhost:8000`, and some other links that were broken.

Fixes: #23