Add Rubyric example and brief description of the supported options
A new example using a **config.yaml** file was added to the Manual.
This new exercise will allow the users of this manual to be aware of
the additional options that can be set for the Rubyric exercises. A
short paragraph was also added to describe in some of the options in
more detail.