Separate internal and external enrollment questionnaires
A+ has used Sphinx 1.6 in the past. New Sphinx 4.1 and
Docutils 0.17 produce an error "circular inclusion" if two
questionnaires in the same chapter include the same RST file.
The enrollment questionnaire template uses include, thus it is
necessary to separate the internal and external questionnaires
to different chapters in order to avoid the error in "include".